chore(format): format fixed after merge
