+++ tsconfig.json
@@ -5,92 +5,29 @@
   },
   "files": [],
   "references": [
-    {
-      "path": "packages/types"
-    },
-    {
-      "path": "packages/errors"
-    },
-    {
-      "path": "packages/config"
-    },
-    {
-      "path": "packages/logger"
-    },
-    {
-      "path": "packages/mongo"
-    },
-    {
-      "path": "packages/sqlite"
-    },
-    {
-      "path": "packages/redis"
-    },
-    {
-      "path": "packages/queue"
-    },
-    {
-      "path": "packages/llm"
-    },
-    {
-      "path": "packages/ingest-github"
-    },
-    {
-      "path": "packages/ingest-business-context"
-    },
-    {
-      "path": "packages/cli"
-    },
-    {
-      "path": "packages/server"
-    },
-    {
-      "path": "packages/neo4j"
-    },
-    {
-      "path": "packages/mcp"
-    },
-    {
-      "path": "packages/db-core"
-    },
-    {
-      "path": "packages/graph-core"
-    },
-    {
-      "path": "packages/queue-core"
-    },
-    {
-      "path": "packages/queue-bullmq"
-    },
-    {
-      "path": "packages/queue-honker"
-    },
-    {
-      "path": "packages/db"
-    },
-    {
-      "path": "packages/graph-db"
-    }
     { "path": "packages/types" },
     { "path": "packages/errors" },
     { "path": "packages/config" },
-    { "path": "packages/path-migration" },
     { "path": "packages/logger" },
     { "path": "packages/mongo" },
     { "path": "packages/sqlite" },
     { "path": "packages/redis" },
+    { "path": "packages/neo4j" },
+    { "path": "packages/ladybug" },
+    { "path": "packages/db-core" },
+    { "path": "packages/db" },
+    { "path": "packages/graph-core" },
+    { "path": "packages/graph-db" },
+    { "path": "packages/queue-core" },
+    { "path": "packages/queue-bullmq" },
+    { "path": "packages/queue-honker" },
     { "path": "packages/queue" },
     { "path": "packages/llm" },
+    { "path": "packages/path-migration" },
     { "path": "packages/ingest-github" },
     { "path": "packages/ingest-business-context" },
-    { "path": "packages/cli" },
-    { "path": "packages/server" },
-    { "path": "packages/neo4j" },
     { "path": "packages/mcp" },
-    { "path": "packages/db-core" },
-    { "path": "packages/graph-core" },
-    { "path": "packages/db" },
-    { "path": "packages/graph-db" },
-    { "path": "packages/ladybug" }
+    { "path": "packages/cli" },
+    { "path": "packages/server" }
   ]
 }
